refactor: update application to use BigDecimal for financial calculations and remove unused database configuration classes
- Replaced Double with BigDecimal in various controllers and services for improved precision in financial operations.
- Removed manual DataSource configuration classes (DatabaseConfig and RenderDatabaseConfig) in favor of Spring Boot's auto-configuration.
- Cleaned up pom.xml by removing unnecessary dependencies.
- Enhanced CORS and security configurations for better API access control.

+++ src/main/java/com/bypass/bypasstransers/controller/AnalysisController.java
@@ -1,13 +1,15 @@
 package com.bypass.bypasstransers.controller;
 
 import com.bypass.bypasstransers.service.AnalysisService;
+import java.math.BigDecimal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
- *
+ * REST API for profit analysis
+ * 
  * @author Vickeller.01
  */
 @RestController
@@ -18,7 +20,7 @@ public class AnalysisController {
     private AnalysisService service;
 
     @GetMapping("/profit")
-    public double profit() {
+    public BigDecimal profit() {
         return service.totalProfit();
     }
 }
