+++ src/main/java/com/bypass/bypasstransers/repository/ExpenditureRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -18,7 +19,7 @@ public interface ExpenditureRepository extends JpaRepository<Expenditure, Long> 
     List<Expenditure> findByCategoryAndDateBetweenOrderByDateDesc(String category, LocalDate startDate, LocalDate endDate);
     
     @Query("SELECT SUM(e.amount) FROM Expenditure e WHERE e.date BETWEEN :startDate AND :endDate")
-    Double getTotalExpenditureForPeriod(@Param("startDate") LocalDate startDate, @Param("endDate") LocalDate endDate);
+    BigDecimal getTotalExpenditureForPeriod(@Param("startDate") LocalDate startDate, @Param("endDate") LocalDate endDate);
     
     @Query("SELECT e.category, SUM(e.amount) FROM Expenditure e WHERE e.date BETWEEN :startDate AND :endDate GROUP BY e.category")
     List<Object[]> getExpenditureByCategoryForPeriod(@Param("startDate") LocalDate startDate, @Param("endDate") LocalDate endDate);
